+++ src/pocketmine/level/generator/end/biome/EndBiome.php
@@ -22,7 +22,7 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\level\generator\end;
+namespace pocketmine\level\generator\end\biome;
 
 use pocketmine\level\generator\biome\Biome;
