+++ src/pocketmine/level/generator/end/populator/EndPillar.php
@@ -62,7 +62,7 @@ class EndPillar extends Populator{
 						for($r = 0.5; $r < 5; $r += 0.5){
 							$nd = 360 / (2 * pi() * $r);
 							for($d = 0; $d < 360; $d += $nd){
-								$level->setBlockIdAt($x + (cos(deg2rad($d)) * $r), $ny, $z + (sin(deg2rad($d)) * $r), Block::OBSIDIAN);
+								$level->setBlockIdAt( (int) $x + (cos(deg2rad($d)) * $r), $ny, (int) $z + (sin(deg2rad($d)) * $r), Block::OBSIDIAN);
 							}
 						}
 					}
